drf_ecom/src/requirements.txt
django-filter==25.2
pip-chill==1.0.5
psycopg2==2.9.12
python-dotenv==1.2.2
